Fix DNS request id in async implementation
The RFC require a 16-bit value and the underlying DNS library enforces
value in the range 0-65535.
In particular using a short to store the current id will fail once we
exceed 32767 (overflow goes to -32768 which is not allowed by the DNS
library).
Instead store the value in an int, and use modulo to restart from 0 once
we overflow 16-bit.