Ignore records that do not start with spf version
According to the RFC 7208 (section 4.5) a SPF record must start with
the version. Thus any leading space would make it a non-SPF record.
This was already correctly handled in the SPF record parser, but not in
the retriever which filters SPF records before parsing/processing them.
This could lead to returning a permerror because such badly formatted
records were taken into account.

1.0.2.AWL.2